Working on task 3

+++ 0x0F-function_pointers/3-op_functions.c
@@ -52,7 +52,7 @@ int op_div(int a, int b)
 		printf("Error\n");
 		exit(100);
 	}
-	return ((int)(a / b));
+	return (a / b);
 }
 
 /**
